+++ R/providers.R
@@ -71,7 +71,7 @@ oauth_provider_oidc <- function(
     userinfo_url = userinfo_url,
     introspection_url = introspection_url,
     issuer = base_url,
-    issuer_thus_oidc = TRUE,
+    infer_oidc_from_issuer = TRUE,
     use_nonce = use_nonce,
     id_token_validation = id_token_validation,
     token_auth_style = token_auth_style,
@@ -164,7 +164,7 @@ oauth_provider_google <- function(name = "google") {
     revocation_url = "https://oauth2.googleapis.com/revoke",
     userinfo_url = "https://openidconnect.googleapis.com/v1/userinfo",
     issuer = "https://accounts.google.com",
-    issuer_thus_oidc = TRUE,
+    infer_oidc_from_issuer = TRUE,
 
     use_nonce = TRUE,
 
@@ -288,7 +288,7 @@ oauth_provider_microsoft <- function(
     introspection_url = NA_character_,
 
     issuer = issuer,
-    issuer_thus_oidc = TRUE,
+    infer_oidc_from_issuer = TRUE,
     issuer_match = issuer_match,
 
     use_nonce = isTRUE(id_token_validation),